split ModInfoDialogTab
